refactor: move logic into new lib/tally.ts file
